drag card and column sort, existing tag add/remove

+++ database/migrations/2019_09_30_095952_remove_nav_id_from_tasks_table.php
@@ -12,13 +12,19 @@ class RemoveNavIdFromTasksTable extends Migration
      * @return void
      */
     public function up()
-    {
-        Schema::table('task_lists', function (Blueprint $table) {
-            $table->dropColumn('nav_id');
-        });
-        Schema::table('action_logs', function (Blueprint $table) {
-            $table->dropColumn('nav_id');
-        });
+    {   
+        if (Schema::hasColumn('task_lists', 'nav_id'))
+        {
+            Schema::table('task_lists', function (Blueprint $table) {
+                $table->dropColumn('nav_id');
+            });
+        }
+        if (Schema::hasColumn('action_logs', 'nav_id'))
+        {
+            Schema::table('action_logs', function (Blueprint $table) {
+                $table->dropColumn('nav_id');
+            });
+        }
     }
 
     /**
